feat: added new border style types
